actually combine poster checks.
