+++ assets/src/edit-story/components/checklist/checklistContent/priorityChecks.js
@@ -28,9 +28,8 @@ import { ISSUE_TYPES } from '../constants';
 import PublisherLogoSize from '../checks/publisherLogoSize';
 import StoryMissingExcerpt from '../checks/storyMissingExerpt';
 import StoryMissingTitle from '../checks/storyMissingTitle';
-import StoryPosterAspectRatio from '../checks/storyPosterAspectRatio';
+import StoryPosterSize from '../checks/storyPosterSize';
 import { StoryPosterAttached } from '../checks/storyPosterAttached';
-import StoryPosterPortraitSize from '../checks/storyPosterPortraitSize';
 import StoryTitleLength from '../checks/storyTitleLength';
 import VideoElementMissingPoster from '../checks/videoElementMissingPoster';
 import { ChecklistCategoryProvider, useCategoryCount } from '../countContext';
@@ -74,8 +73,7 @@ export function PriorityChecks({
         <StoryTitleLength />
         <StoryMissingExcerpt />
         <StoryPosterAttached />
-        <StoryPosterPortraitSize />
-        <StoryPosterAspectRatio />
+        <StoryPosterSize />
         <PublisherLogoSize />
         <VideoElementMissingPoster />
         {isTranscodingEnabled && <VideoOptimization />}
